Move feature out of dev-only
Make sendgrid feature declaration accessible to prod
