+++ src/Settings/Features.php
@@ -10,12 +10,12 @@ use P4\MasterTheme\Features\Dev\CoreBlockPatterns;
 use P4\MasterTheme\Features\Dev\DisableDataSync;
 use P4\MasterTheme\Features\Dev\ListingPageGridView;
 use P4\MasterTheme\Features\Dev\WPTemplateEditor;
-use P4\MasterTheme\Features\Dev\Sendgrid;
 use P4\MasterTheme\Features\EngagingNetworks;
 use P4\MasterTheme\Features\LazyYoutubePlayer;
 use P4\MasterTheme\Features\PurgeOnFeatureChanges;
 use P4\MasterTheme\Features\RedirectRedirectPages;
 use P4\MasterTheme\Features\Planet4Blocks;
+use P4\MasterTheme\Features\Sendgrid;
 use P4\MasterTheme\Loader;
 use P4\MasterTheme\Settings;
 use CMB2;
